Notes_pipeline
CRISPR- arrays

Identifying the CRISPR arrays in a set of genomes

1.-Download genomes with python Script

CRISPR-Finder and Piler CR use different approaches. 

PILER is the Parsimonous inference of a library of elementary repeats

The input can be a genome or fragments of a genome.
There are 4 methods of search: Dispersed family, tandem array, pseudo satellite and terminal repeat. 
It requires an aligner, in this case is PALS. 2. Regions aligned by a hit are called an image. Copy count is the number of images that cover the base. Aligner is not available anymore

CRISPR-Finder

Already ran a test of 20 genomes from the File of pathogenic strains

Getting the phylogenetic origin
-NOTE: (Some filtering will be useful, following Koonin´s approach- ORPHAN CRISPR, those ones that are not complete and getting the ones that at least in other bacteria are complete)

The blast should be stringent given the size of the spacers


---
In parallel, Checking Koonin´s resources
There is a file in his server folder "CRISPRclass19"

Koonin